fix bug $this->auxPDO is NULL and load pages without Backend

+++ Modelo/clsCarritoCRUD.php
@@ -17,6 +17,7 @@ class clsCarritoCRUD {
     public function Crear($obj){
         $resultado = false;
         try{
+            if($this->auxPDO == NULL) return $resultado;
             $consulta = "INSERT INTO `agregacarrito` (PRO_ID,USU_ID,CAR_CANTIDAD) VALUES (?,?,?)";
             $this->auxPDO->prepare($consulta)->execute(array($obj->proid,$obj->usuid,$obj->cantidad));
             $resultado = true;
@@ -30,6 +31,7 @@ class clsCarritoCRUD {
     public function Editar($obj){
         $resultado = false;
          try{
+            if($this->auxPDO == NULL) return $resultado;
             $consulta = "UPDATE `agregacarrito` SET CAR_CANTIDAD=? WHERE USU_ID = ? AND PRO_ID =?";
             $this->auxPDO->prepare($consulta)->execute(array($obj->cantidad, $obj->usuid, $obj->proid));
             $resultado = true;
@@ -43,6 +45,7 @@ class clsCarritoCRUD {
     public function Eliminar($obj){
         $resultado = false;
         try{
+            if($this->auxPDO == NULL) return $resultado;
             $consulta = "DELETE FROM `agregacarrito` WHERE USU_ID = ? AND PRO_ID =?";
             $this->auxPDO->prepare($consulta)->execute(array($obj->usuid,$obj->proid));
             $resultado = true;
@@ -56,6 +59,7 @@ class clsCarritoCRUD {
     public function ObtenerProductos($id){
         $resultado = array();
         try{
+            if($this->auxPDO == NULL) return $resultado;
             $consulta = $this->auxPDO->prepare("SELECT P.PRO_ID, CC.USU_ID, P.PRO_NOMBRE,P.PRO_PRECIO,P.PRO_IMAGEN,P.PRO_DESCRIPCION,CC.CAR_CANTIDAD,CP.CATPRO_NOMBRE FROM `agregacarrito` CC INNER JOIN `producto` P ON CC.PRO_ID = P.PRO_ID INNER JOIN `categoriaproducto` CP ON P.CATPRO_ID = CP.CATPRO_ID WHERE USU_ID = ?");
             $consulta->execute(array($id));
             foreach ($consulta->fetchALL(PDO::FETCH_OBJ) as $obj){
@@ -80,6 +84,7 @@ class clsCarritoCRUD {
     public function obtenerCantidad($usuid, $proid){
         $cantidad = 0;
         try{
+            if($this->auxPDO == NULL) return $cantidad;
             $consulta = $this->auxPDO->prepare("SELECT sum(CAR_CANTIDAD) as suma FROM `agregacarrito` WHERE PRO_ID = ? AND USU_ID = ?");
             $consulta->execute(array($proid, $usuid));
             if(($cantidad = $consulta->fetchALL()) != NULL){
